Short refactor
