+++ src/main/java/com/mx/development/said/olano/aggregator/NewsCSVLineAggregator.java
@@ -1,0 +1,18 @@
+package com.mx.development.said.olano.aggregator;
+
+import com.mx.development.said.olano.entity.NewsProcess;
+import com.mx.development.said.olano.extractor.NewsCSVFieldExtractor;
+import org.springframework.batch.item.file.transform.DelimitedLineAggregator;
+
+public class NewsCSVLineAggregator extends DelimitedLineAggregator<NewsProcess> {
+
+    public NewsCSVLineAggregator(){
+        setDelimiter(",");
+        setFieldExtractor(new NewsCSVFieldExtractor());
+    }
+
+    public String aggregate(NewsProcess item) {
+        return super.aggregate(item);
+    }
+
+}
