+++ src/main/java/com/mx/development/said/olano/commons/Constants.java
@@ -1,0 +1,8 @@
+package com.mx.development.said.olano.commons;
+
+public class Constants {
+    public static final String CLEAN_UP_WORKING_DIRECTORY_OS_JOB ="cleanUpWorkingDirectoryOSJob";
+    public static final String RETRIEVE_NEWS_JOB ="retrieveNewsJob";
+    public static final String PROCESS_NEWS_JOB ="processNewsJob";
+    public static final String JOB_ID = "JobID";
+}
