--- src/main/java/pe/edu/galaxy/training/java/sb/spring/batch/appspring/batchdemocollectionv1/configuration/DataSourceConfig.java
@@ -1,45 +1,0 @@
-package pe.edu.galaxy.training.java.sb.spring.batch.appspring.batchdemocollectionv1.configuration;
-
-import org.springframework.boot.autoconfigure.quartz.QuartzDataSource;
-import org.springframework.boot.autoconfigure.quartz.QuartzTransactionManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.jdbc.datasource.init.DataSourceInitializer;
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-
-import javax.sql.DataSource;
-
-@Configuration
-public class DataSourceConfig {
-
-	@QuartzDataSource
-	private DataSource dataSource;
-
-	public DataSourceConfig(DataSource dataSource) {
-		this.dataSource = dataSource;
-	}
-
-	@Bean
-	@QuartzTransactionManager
-	public DataSourceTransactionManager transactionManager() {
-		return new DataSourceTransactionManager(dataSource);
-	}
-
-	@Bean
-	public DataSourceInitializer databasePopulator() {
-		
-		ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
-		populator.addScript(new ClassPathResource("org/springframework/batch/core/schema-h2.sql"));
-		populator.setContinueOnError(false);
-		populator.setIgnoreFailedDrops(false);
-		
-		DataSourceInitializer dataSourceInitializer = new DataSourceInitializer();
-		dataSourceInitializer.setDataSource(dataSource);
-		
-		dataSourceInitializer.setDatabasePopulator(populator);
-		
-		return dataSourceInitializer;
-	}
-}
