+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8082
+server.port=8080
 spring.application.name=app-spring.batch-demo-collection-v1
 
 spring.datasource.url=jdbc:h2:mem:testdb
